supa down and dirty implementation.
tl;dr: Rough City

This does not perform any proper checks for Gravity Form capabilities, has some bugs when adding new fields it will auto set the custom post value as the previous field.

Testing out the singleton pattern, not sure if there are going to be any issues with it or not.

Poorly commented

Poorly names all over the place.  I don't even know what to call this at the moment so the naming conventions are kind of all over the place.